Apply suggestions from code review

+++ src/map/guild.cpp
@@ -1760,7 +1760,7 @@ int32 guild_skillupack(int32 guild_id,uint16 skill_id,uint32 account_id) {
 	if (g == nullptr || idx == -1)
 		return 0;
 	if (sd != nullptr) {
-		int32 lv = g->guild.skill[idx].lv;
+		uint16 lv = g->guild.skill[idx].lv;
 		int32 range = skill_get_range(skill_id, lv);
 		clif_skillup( *sd, skill_id,lv, range, true );
